backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
beautifulsoup4==4.13.4
Flask-Session==0.8.0
gunicorn==23.0.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
